Moved CellWorldDistributedState int impl package of game model
