+++ client/src/main/scala/it/cwmp/client/model/game/impl/CellWorldDistributedState.scala
@@ -1,12 +1,12 @@
-package it.cwmp.client.model
+package it.cwmp.client.model.game.impl
 
 import akka.actor.Actor.Receive
 import akka.actor.ActorRef
 import akka.cluster.Cluster
-import akka.cluster.ddata.{LWWRegister, Replicator}
 import akka.cluster.ddata.Replicator.{Update, WriteLocal}
+import akka.cluster.ddata.{LWWRegister, Replicator}
+import it.cwmp.client.model.DistributedState
 import it.cwmp.client.model.DistributedState.UpdateState
-import it.cwmp.client.model.game.impl.CellWorld
 
 /**
   * Distributed representation of the world and of his behaviours.
